+++ test/OrchardCore.Tests/Modules/OrchardCore.Media/MediaEventTests.cs
@@ -76,7 +76,7 @@ public class MediaEventTests
 
 public class TestMediaEventHandler : IMediaCreatingEventHandler
 {
-    public int Priority => 0;
+    public double Priority => 0;
 
     public async Task<Stream> MediaCreatingAsync(MediaCreatingContext context, Stream inputStream)
     {
